refactor: separate router and routes
